Crate swiss_qrust

Generate Swiss QR-bill payment slips (ISO 20022 / SIX Implementation Guidelines) as PDF or SVG.

The typical flow is: parse an InputBill from TOML or JSON, convert it to a validated BillData via TryFrom, then render it.

use swiss_qrust::{BillData, InputBill, Language};

let toml = r#"
iban = "CH93 0076 2011 6238 5295 7"
currency = "CHF"
amount = "199.95"

[creditor_address]
name = "Robert Schneider AG"
street = "Rue du Lac"
house_num = "1268"
plz = "2501"
city = "Biel"
country = "CH"
"#;

let input: InputBill = toml::from_str(toml)?;
let bill = BillData::try_from(input)?;

let svg = swiss_qrust::svg::render_bill_to_svg(&bill, Language::De)?;
assert!(svg.starts_with("<svg"));

For a PDF instead of SVG, use pdf::render_bill_to_pdf (returns the raw bytes) or pdf::create_pdf (writes straight to a file). Both render engines consume the same BillData - there is nothing PDF-specific or SVG-specific about validating or building the bill itself.

Re-exports

pub use bill::*;

pub use language::*;

pub use generated::*;

pub use support::*;

pub use render::layout::*;

pub use render::engines::*;

pub use render::types::*;

pub use input::*;

Modules

additional_attributes

Documentation for Additional Attributes

bill

constants

de

Deserialize JSON data to a Rust data structure.

error

When serializing or deserializing JSON goes wrong.

generated

input

language

map

A map of String to serde_json::Value.

render

ser

Serialize a Rust data structure into JSON data.

support

value

The Value enum, a loosely typed way of representing any valid JSON value.

Macros

json

Construct a serde_json::Value from a JSON literal.

label

Exports a macro for the above

Structs

CountryIter

An iterator over the variants of Country

CountryMeta

Deserializer

A structure that deserializes JSON into Rust values.

Map

Represents a JSON key/value type.

Number

Represents a JSON number, whether integer or floating point.

Serializer

A structure for serializing Rust values into JSON.

StreamDeserializer

Iterator that deserializes a stream into multiple JSON values.

Enums

Country

CountryParseError

ParseError

The ParseError enum is a collection of all the possible reasons an enum can fail to parse from a string.

Value

Represents any valid JSON value.

Constants

CORNER_MARKS_AMOUNT_POLYGONS

CORNER_MARKS_AMOUNT_POLYLINES

CORNER_MARKS_AMOUNT_RECTS

CORNER_MARKS_AMOUNT_VIEWBOX

CORNER_MARKS_PAYABLE_BY_POLYGONS

CORNER_MARKS_PAYABLE_BY_POLYLINES

CORNER_MARKS_PAYABLE_BY_RECTS

CORNER_MARKS_PAYABLE_BY_VIEWBOX

CROSS_POLYGONS

CROSS_POLYLINES

CROSS_RECTS

CROSS_VIEWBOX

Traits

AsStaticRefDeprecated

A cheap reference-to-reference conversion. Used to convert a value to a reference value with 'static lifetime within generic code.

EnumCount

A trait for capturing the number of variants in Enum. This trait can be autoderived by strum_macros.

EnumMessage

Associates additional pieces of information with an Enum. This can be autoimplemented by deriving EnumMessage and annotating your variants with #[strum(message="...")].

EnumProperty

EnumProperty is a trait that makes it possible to store additional information with enum variants. This trait is designed to be used with the macro of the same name in the strum_macros crate. Currently, the string, integer and bool literals are supported in attributes.

IntoDiscriminant

A trait for retrieving the enum generated by EnumDiscriminants from an associated Type on the original enumeration. This trait can be autoderived by strum_macros.

IntoEnumIterator

This trait designates that an Enum can be iterated over. It can be auto generated using the EnumIter derive macro.

VariantArray

A trait for retrieving a static array containing all the variants in an Enum. This trait can be autoderived by strum_macros. For derived usage, all the variants in the enumerator need to be unit-types, which means you can’t autoderive enums with inner data in one or more variants. Consider using it alongside EnumDiscriminants if you require inner data but still want to have an static array of variants.

VariantIterator

VariantMetadata

VariantNames

A trait for retrieving the names of each variant in Enum. This trait can be autoderived by strum_macros.

Functions

from_reader

Deserialize an instance of type T from an I/O stream of JSON.

from_slice

Deserialize an instance of type T from bytes of JSON text.

from_str

Deserialize an instance of type T from a string of JSON text.

from_value

Interpret a serde_json::Value as an instance of type T.

parse_bill_data

Parses content as an InputBill, choosing the format by extension.

to_string

Serialize the given data structure as a String of JSON.

to_string_pretty

Serialize the given data structure as a pretty-printed String of JSON.

to_value

Convert a T into serde_json::Value which is an enum that can represent any valid JSON data.

to_vec

Serialize the given data structure as a JSON byte vector.

to_vec_pretty

Serialize the given data structure as a pretty-printed JSON byte vector.

to_writer

Serialize the given data structure as JSON into the I/O stream.

to_writer_pretty

Serialize the given data structure as pretty-printed JSON into the I/O stream.

Derive Macros

AsRefStr

Converts enum variants to &'a str, where 'a is the lifetime of the input enum reference.

Display

Converts enum variants to strings.

EnumCount

Add a constant usize equal to the number of variants.

EnumDiscriminants

Generate a new type with only the discriminant names.

EnumIs

Generated is_*() methods for each variant. E.g. Color.is_red().

EnumIter

Creates a new type that iterates over the variants of an enum.

EnumMessage

Add a verbose message to an enum variant.

EnumProperty

Add custom properties to enum variants.

EnumString

Converts strings to enum variants based on their name.

EnumTryAs

Generated try_as_*() methods for all tuple-style variants. E.g. Message.try_as_write().

FromRepr

Add a function to enum that allows accessing variants by its discriminant

IntoStaticStr

Implements From<MyEnum> for &'static str on an enum.

VariantArray

Adds a 'static slice with all of the Enum’s variants.

VariantNames

Implements Strum::VariantNames which adds an associated constant VARIANTS which is a 'static slice of discriminant names.