swiss_qrust::render::layout

Module geometry

Structs

Baseline

Mm

Conversion structs. We use const fns to do compile time evaluation and structs to facilitate type checking up front.

Pt

QRBillLayoutRect

Constants

A4_PORTRAIT_HEIGHT

A4 Height in mm

A4_PORTRAIT_WIDTH

A4 Width in mm

MM_PER_PT

Constants to convert mms to pts and vice versa.

PT_PER_MM

QR_BILL_HEIGHT

Now the height of the slip itself

QR_BILL_PC_WIDTH

QR_BILL_RC_WIDTH

QR_BILL_SLIP_WIDTH

Width of a QR Slip is incidentally the same as the whole paper slip 🤣

QR_BILL_WITH_ALL_WIDTH

Then the width with scissors symbol and all… Surprise

QR_BILL_WITH_HORI_LINE_HEIGHT

Takes 5 mm more with all the scissors assets.

QR_CODE_OVERALL_WIDTH

QR Width in mm